tee: support non-clone types
This commit makes TeeCore able to handle types that don't implement
`Clone` by dynamicaly changing the push implementation depending on the
number of pushers. This is enforced statically using the following
mechanism:

A `TeeCore`/`TeeHelper` can be initialized for any `T, D` and they start
out with zero pushers in the list of pushers.

A `TeeHelper` represents the capability to add a single pusher to the
list of pushers. This is achieved by making `TeeHelper::add_pusher`
consume `self`.

The capability to add a single pusher can be cloned in onder to add more
pushers but crucially this is only implemented when `T` and `D` are
cloneable. When a `TeeHelper` gets cloned it also dynamically switches
the `push` implementation of `TeeCore` to one that will clone the values
for each destination pusher.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>